support grep and filter

+++ xiongkun/plugin/pythonx/Xiongkun/rpc_server/vimrpc/remote_fs.py
@@ -2,20 +2,23 @@ import os
 import sys
 import json
 import time
-from decorator import *
-from log import log
-from functions import KillablePool
-
+from .decorator import *
+from .functions import KillablePool
+import os.path as osp
 
 class RemoteFS:
     @server_function
     def fetch(self, path):
+        if not osp.isfile(path):
+            return f"Error: no such files: `{path}`"
         with open(path, "r") as fp: 
             lines = fp.readlines()
         return "".join(lines)
 
     @server_function
     def store(self, path, content):
+        if not osp.isfile(path):
+            return f"Error: no such files: `{path}`"
         try:
             with open(path, "w") as fp: 
                 fp.write(content)
